talk_to_docs/requirements.txt
streamlit>=1.35.0
openai>=1.35.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
numpy>=1.26.4
python-dotenv>=1.0.1
PyPDF2>=3.0.1
python-docx>=1.1.2
tenacity>=8.4.2
